Fix Stream enabling code in example TS scripts
Added  IDeviceStreamApiImpl interface and changed example device scripts to use it to make it less likely that script writers introduce stream.enabled related bugs by writing code like:   if (stream && stream.enabled) {...} which is always true because enabled is a setting object containing a boolean, not a boolean!
ProxyDevice.initializeSettings() now properly handles the cases where the number of streams in the settings does not match the number in the current hardware.